Removed the external server for executor
