+++ charts/fission-all/templates/deployment.yaml
@@ -226,7 +226,7 @@ spec:
         image: "{{ .Values.image }}:{{ .Values.imageTag }}"
         imagePullPolicy: {{ .Values.pullPolicy }}
         command: ["/fission-bundle"]
-        args: ["--routerPort", "8888", "--executorUrl", "http://executor.{{ .Release.Namespace }}"]
+        args: ["--routerPort", "8888", "--poolmgrUrl", "http://poolmgr.{{ .Release.Namespace }}"]
       serviceAccount: fission-svc
 
 ---
@@ -273,50 +273,7 @@ spec:
         - name: RUNTIME_IMAGE_PULL_POLICY
           value: "{{ .Values.pullPolicy }}"
       serviceAccount: fission-svc
----
-apiVersion: v1
-kind: Service
-metadata:
-  name: executor
-  labels:
-    svc: executor
-    chart: "{{ .Chart.Name }}-{{ .Chart.Version }}"
-spec:
-  type: ClusterIP
-  ports:
-  - port: 80
-    targetPort: 8890
-  selector:
-    svc: executor
 
----
-apiVersion: extensions/v1beta1
-kind: Deployment
-metadata:
-  name: executor
-  labels:
-    chart: "{{ .Chart.Name }}-{{ .Chart.Version }}"
-spec:
-  replicas: 1
-  template:
-    metadata:
-      labels:
-        svc: executor
-    spec:
-      containers:
-      - name: executor
-        image: "{{ .Values.image }}:{{ .Values.imageTag }}"
-        imagePullPolicy: {{ .Values.pullPolicy }}
-        command: ["/fission-bundle"]
-        args: ["--executorPort", "8890", "--poolmgrUrl", "http://poolmgr.{{ .Release.Namespace }}"]
-        env:
-        - name: FETCHER_IMAGE
-          value: "{{ .Values.fetcherImage }}:{{ .Values.fetcherImageTag }}"
-        - name: FETCHER_IMAGE_PULL_POLICY
-          value: "{{ .Values.pullPolicy }}"
-        - name: RUNTIME_IMAGE_PULL_POLICY
-          value: "{{ .Values.pullPolicy }}"
-      serviceAccount: fission-svc
 ---
 apiVersion: v1
 kind: Service
