--- executor/client/client.go
@@ -1,80 +1,0 @@
-/*
-Copyright 2016 The Fission Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package client
-
-import (
-	"bytes"
-	"encoding/json"
-	"io/ioutil"
-	"net/http"
-	"strings"
-
-	"github.com/fission/fission"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-)
-
-type Client struct {
-	executorUrl string
-}
-
-func MakeClient(executorURL string) *Client {
-	c := &Client{
-		executorUrl: strings.TrimSuffix(executorURL, "/"),
-	}
-	return c
-}
-
-func (c *Client) GetServiceForFunction(metadata *metav1.ObjectMeta) (string, error) {
-	executorURL := c.executorUrl + "/v2/getServiceForFunction"
-
-	body, err := json.Marshal(metadata)
-	if err != nil {
-		return "", err
-	}
-
-	resp, err := http.Post(executorURL, "application/json", bytes.NewReader(body))
-	if err != nil {
-		return "", err
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode != 200 {
-		return "", fission.MakeErrorFromHTTP(resp)
-	}
-
-	svcName, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		return "", err
-	}
-
-	return string(svcName), nil
-
-}
-
-func (c *Client) TapService(serviceUrl string) error {
-	executorURL := c.executorUrl + "/v2/tapService"
-
-	resp, err := http.Post(executorURL, "application/octet-stream", bytes.NewReader([]byte(serviceUrl)))
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-	if resp.StatusCode != 200 {
-		return fission.MakeErrorFromHTTP(resp)
-	}
-	return nil
-}
